+++ src/frontend/apps/drive/src/features/explorer/components/trash/utils.tsx
@@ -1,4 +1,8 @@
-import { useModals } from "@gouvfr-lasuite/cunningham-react";
+import {
+  ModalSize,
+  useModals,
+  VariantType,
+} from "@gouvfr-lasuite/cunningham-react";
 import i18n from "@/features/i18n/initI18n";
 
 export const messageModalTrashNavigate = (
@@ -7,6 +11,8 @@ export const messageModalTrashNavigate = (
 ) => {
   const key = isFile ? "modal_file" : "modal_folder";
   modals.messageModal({
+    messageType: VariantType.INFO,
+    size: ModalSize.MEDIUM,
     title: i18n.t(`explorer.trash.navigate.${key}.title`),
     children: (
       <div className="clr-greyscale-600">